refactor(gamestate): promote area tracker to Mithril.GameState (#454)
Move AreaTransitionParser, PlayerAreaTracker, and the extracted
AreaTransitionEvent out of Gandalf into Mithril.GameState.Areas — shared
live game-state, mirroring the existing Quest/Inventory/Session services.
Registered once in AddMithrilGameState(); Gandalf now injects the tracker
instead of registering its own. No behaviour change for Gandalf.

This is the foundation for #454: Legolas needs the same LOADING LEVEL
area signal to drive per-area survey-projection calibration, and the
design of record calls for one shared tracker (no duplicated regex).

- AreaTransitionEvent extracted from Gandalf LootEvents.cs into
  Areas/Parsing/AreaEvents.cs (loot events stay put).
- Diagnostics category "Gandalf.Area" -> "GameState.Area" (accurate
  owner); doc cref to Gandalf-only LootSource generalised.
- Parser/tracker tests relocated to Mithril.GameState.Tests/Areas;
  Gandalf LootSourceTests usings updated. Full solution builds clean;
  29 area tests + 37 Gandalf LootSource tests green.
